Use 32-bit WAV files (floats) instead of 16-bit (shorts)
* Added frequency "boosting" or "mixing"
* Increased default sampling rate to 88,200 samples/second
* Added event hooks for outside observers

+++ Core/AudioAnalysis/SamplingResult.cs
@@ -4,7 +4,7 @@ namespace Core.AudioAnalysis
 {
     public class SamplingResult
     {
-        public IList<short> Samples { get; set; }
+        public IList<float> Samples { get; set; }
         public int SampleRate { get; set; }
         public int BitsPerSample { get; set; }
         public double FileLengthInMilliseconds { get; set; }
